+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/Luck.java
@@ -13,8 +13,8 @@ import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.item.enchantment.LootBonusEnchantment;
 
 public class Luck extends LootBonusEnchantment implements IEnchantmentTooltip {
-    public static final TagKey<Item> ACCEPTS_LUCK = ITRItemTagsProvider.create("enchantments/accepts_luck");
-    static final EnchantmentCategory CATEGORY = EnchantmentCategory.create("luck", item -> item.builtInRegistryHolder().is(ACCEPTS_LUCK));
+    public static final TagKey<Item> ACCEPTS_ENCHANTMENT = ITRItemTagsProvider.create("enchanting/accepts_luck");
+    static final EnchantmentCategory CATEGORY = EnchantmentCategory.create("luck", item -> item.builtInRegistryHolder().is(ACCEPTS_ENCHANTMENT));
     public Luck() {
         super(Rarity.RARE, CATEGORY, EquipmentSlot.MAINHAND);
     }