+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/damage/Sharpness.java
@@ -1,11 +1,10 @@
 package insane96mcp.iguanatweaksreborn.module.experience.enchantments.enchantment.damage;
 
-import insane96mcp.iguanatweaksreborn.module.experience.enchantments.EnchantmentsFeature;
 import net.minecraft.world.entity.EquipmentSlot;
 
 public class Sharpness extends BonusDamageEnchantment {
     public Sharpness() {
-        super(Rarity.UNCOMMON, EnchantmentsFeature.WEAPONS_CATEGORY, new EquipmentSlot[]{ EquipmentSlot.MAINHAND });
+        super(Rarity.UNCOMMON, new EquipmentSlot[]{ EquipmentSlot.MAINHAND });
     }
 
     @Override