+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/ThornsEnchantmentMixin.java
@@ -36,11 +36,11 @@ public abstract class ThornsEnchantmentMixin {
         if (!EnchantmentsFeature.isThornsOverhaul())
             return;
 
-        //lvl~lvl+lvl*2
-        //1~3
-        //2~6
-        //3~9
-        //4~12
-        cir.setReturnValue(lvl + random.nextInt(lvl * 2 + 1));
+        //lvl~lvl+(lvl-1)
+        //1~1
+        //2~3
+        //3~5
+        //4~7
+        cir.setReturnValue(lvl + random.nextInt(lvl));
     }
 }