+++ changelog.md
@@ -7,6 +7,7 @@
 * Reduced Honey bottle healing
 * Reduced flint and steel durability
 * Fixed hoes having too much Attack Damage
+* Fixed enchantments not scaling with attack cooldown
 
 ## 3.5.1
 * Spawners no longer give loot but give more experience