+++ changelog.md
@@ -10,6 +10,8 @@
   * Item tags: `iguanatweaksreborn:enchanting/accepts_damage_enchantments`, `iguanatweaksreborn:enchanting/accepts_fire_aspect`, `iguanatweaksreborn:enchanting/accepts_knockback`, `iguanatweaksreborn:enchanting/accepts_luck`
 * Fixed hoes having too much Attack Damage
   * With this, attack speed has been reverted to 2.5
+* Fixed repairing enchanted items with non-enchanted ones costing levels
+  * Also can no longer merge broken items
 * Fixed enchantments not scaling with attack cooldown
 
 ## 3.5.1